+++ examples/matrix_operations/matrix_16x16.json
@@ -5,5 +5,6 @@
       "WEIGHT": [16, 16]
     }
   ],
+  "GEMM": ["INPUT", "WEIGHT"],
   "BASE_ADDRESS" : "0000"
 }
