+++ examples/matrix_operations/matrix_20x20.json
@@ -6,8 +6,16 @@
       "ACCUMULATOR": [20, 20]
     }
   ],
+  "GEMM": ["INPUT", "WEIGHT", "ACCUMULATOR"],
   "ALU" : [
-    ["MAX_IMM", [[0,1], 0, 20]]
+    ["ADD", [[0,0], [1,1], 4]],
+    ["SHR_IMM", [0, 2]],
+    ["ADD", [[5,0], [6,1], 4]],
+    ["SHR_IMM", [5, 2]],
+    ["ADD", [[10,0], [11,1], 4]],
+    ["SHR_IMM", [10, 2]],
+    ["ADD", [[15,0], [16,1], 4]],
+    ["SHR_IMM", [15, 2]]
   ],
   "BASE_ADDRESS" : "0000"
 }
