+++ src/compiler/vta_compiler/data_definition/matrix_generator.py
@@ -5,13 +5,13 @@ import numpy as np
 
 # MATRIX GENERATOR FUNCTIONS
 # --------------------------
-def matrix_creation(n_row=16, n_col=16, isInitRandom=True, random_bound=128, dtype=np.int8, onlyPositive=False):
+def matrix_creation(n_row=16, n_col=16, isInitRandom=True, random_bound=0, dtype=np.int8, onlyPositive=False):
     """Create a matrix with dtype values (e.g., int8 or int32)."""
     low_bound = 0 if onlyPositive else -random_bound
     if (isInitRandom): # Random values
         matrix = np.random.randint(low_bound, random_bound - 1, size=(n_row, n_col), dtype=dtype)
-    else: # Init with 0
-        matrix = np.zeros((n_row, n_col), dtype=dtype)
+    else: # Init with same value everywhere
+        matrix = np.full((n_row, n_col), random_bound, dtype=dtype)
     # Return the matrix
     return matrix
