+++ src/compiler/vta_compiler/main_vta_compiler.py
@@ -43,11 +43,12 @@ def main(operations_dict, vta_config_dict, debug=True):
 
     # ---------------------------------------------
     # DATA DEFINITION
-    print(f"\nDEBUG: DATA DEFINITION")
-    A_blocks, A_blocks_col, B_blocks, B_blocks_col, X_blocks, X_blocks_col, Y_blocks, ALU_blocks, C_blocks, C_blocks_col, C_init, alu_operations, idx_to_store, doGemm, doAddMatrix = \
-        DF.data_definition(operations_dict, inp_dtype=inp_dtype, wgt_dtype=wgt_dtype, acc_dtype=acc_dtype,
-                           block_size=block_size, random_bound=random_bound, debug=debug)
-    print(f"\nDEBUG: END DATA DEFINITION \n\n")
+    A_blocks, A_blocks_col, B_blocks, B_blocks_col, \
+        X_blocks, Y_blocks, ALU_blocks, C_blocks, C_init, X_blocks_col, \
+        alu_operations, idx_to_store, \
+        doGemm, doAddMatrix, doAlu = \
+            DF.data_definition(operations_dict, inp_dtype=inp_dtype, wgt_dtype=wgt_dtype, acc_dtype=acc_dtype,
+                               block_size=block_size, random_bound=random_bound, debug=debug)
 
 
     # ---------------------------------------------
@@ -74,13 +75,11 @@ def main(operations_dict, vta_config_dict, debug=True):
         dram_offset = 0x0000
 
 
-    print(f"\nDEBUG: DRAM ALLOCATION")
     # Allocate the object
     base_addresses_list, current_dram_addr = \
         DA.dram_allocation(object_list, base_addr=base_address, block_size=block_size, 
                            inp_dtype=inp_dtype, wgt_dtype=wgt_dtype, acc_dtype=acc_dtype,
                            dram_offset=dram_offset, debug=debug)
-    print(f"\nDEBUG: END DRAM ALLOCATION\n\n")
 
 
     # ---------------------------------------------
@@ -94,36 +93,31 @@ def main(operations_dict, vta_config_dict, debug=True):
         nb_A = 0
         nb_B = 0
     nb_X = len(X_blocks)
-    nb_C = len(C_blocks)
 
     # Select a strategy in case of overfitting
     strategy_selector = 1
 
-    print(f"\nDEBUG: MATRIX PARTITIONING")
     # Apply matrix partitioning (check is overfit then applies selected trategy)
     # => Strategy 1 is the most naive, but works in the most of the cases (strategies 2 to 4 have more constrained assumptions)
     isOverfitting, strategy = \
         MP.matrix_partitioning(nb_A=nb_A, A_blocks_col=A_blocks_col, nb_B=nb_B, B_blocks_col=B_blocks_col, 
-                               nb_X=nb_X, X_blocks_col=X_blocks_col, nb_C=nb_C, C_blocks_col=C_blocks_col,
+                               nb_X=nb_X, X_blocks_col=X_blocks_col,
                                inp_buffer_size=inp_buffer_size, wgt_buffer_size=wgt_buffer_size, 
                                acc_buffer_size=acc_buffer_size, out_buffer_size=out_buffer_size,
                                alu_operations=alu_operations, idx_to_store=idx_to_store,
-                               doAddMatrix=doAddMatrix,
+                               doGemm=doGemm, doAddMatrix=doAddMatrix, doAlu=doAlu,
                                strategy_selector=strategy_selector, block_size=block_size,
                                debug=debug)
-    print(f"\nDEBUG: END MATRIX PARTITIONING \n\n")
 	
 
     # ---------------------------------------------
     # OPERATIONS DEFINITION
 
-    print(f"\nDEBUG: OPERATION DEFINITION")
     insn_buffer, uop_buffer = \
         OP.operations_definition(strategy=strategy, dram_addresses=base_addresses_list,
                                  operations_dict=operations_dict, block_size=block_size, uop_buffer_size=uop_buffer_size,
-                                 A_blocks_col=A_blocks_col, B_blocks_col=B_blocks_col, X_blocks_col=X_blocks_col, C_blocks_col=C_blocks_col,
+                                 A_blocks_col=A_blocks_col, B_blocks_col=B_blocks_col, X_blocks_col=X_blocks_col,
                                  debug=debug)
-    print(f"\nDEBUG: END OPERATION DEFINITION \n\n")
     
     # Update DRAM allocation
     object_list = [("UOP", uop_buffer),
@@ -139,19 +133,23 @@ def main(operations_dict, vta_config_dict, debug=True):
     # ---------------------------------------------
     # DATA BINARISATION
 
-    print(f"\nDEBUG: BINARISATION")
+    # Get the name of the execution
+    if "NAME" in operations_dict:
+        name = operations_dict["NAME"]
+    else:
+        name = ''
     # Setup the output folder (standalone-vta/compiler_output/)
     output_dir = compiler_output_setup()
 
     # MATRICES
     # Define the complete path of the files
-    A_blocks_file_path = filepath_definition(output_dir, 'input.bin')
-    B_blocks_file_path = filepath_definition(output_dir, 'weight.bin')
-    X_blocks_file_path = filepath_definition(output_dir, 'accumulator.bin')
-    Y_blocks_file_path = filepath_definition(output_dir, 'add_accumulator.bin')
-    C_blocks_file_path = filepath_definition(output_dir, 'expected_out.bin')
-    C_init_file_path = filepath_definition(output_dir, 'out_init.bin')
-    ALU_blocks_file_path = filepath_definition(output_dir, 'expected_out_sram.bin')
+    A_blocks_file_path = filepath_definition(output_dir, 'input'+name+'.bin')
+    B_blocks_file_path = filepath_definition(output_dir, 'weight'+name+'.bin')
+    X_blocks_file_path = filepath_definition(output_dir, 'accumulator'+name+'.bin')
+    Y_blocks_file_path = filepath_definition(output_dir, 'add_accumulator'+name+'.bin')
+    C_blocks_file_path = filepath_definition(output_dir, 'expected_out'+name+'.bin')
+    C_init_file_path = filepath_definition(output_dir, 'out_init'+name+'.bin')
+    ALU_blocks_file_path = filepath_definition(output_dir, 'expected_out_sram'+name+'.bin')
 
     # Write A_blocks matrix
     with open(A_blocks_file_path, 'wb') as f:
