+++ src/compiler/vta_compiler/matrix_partitioning/two_matrices_strategies.py
@@ -1,0 +1,8 @@
+# IMPORT PACKAGES
+# ---------------
+import numpy as np
+
+from matrix_partitioning.utils_strategies import *
+
+
+###############################################
